policies client refactor - 2.0.4
