+++ packages/policies-client/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@al/policies",
-  "version": "2.0.3",
+  "version": "2.0.4",
   "license": "MIT",
   "description": "A client for interacting with the Alert Logic Policies Public API",
   "author": {
