+++ packages/policies-client/src/al-policies-client.ts
@@ -6,18 +6,7 @@ import {
   AlDefaultClient,
 } from '@al/core';
 
-interface PolicyFeature {
-  type: string;
-}
-
-export interface Policy {
-  name: string;
-  id: string;
-  product_family: string;
-  features: PolicyFeature[];
-  udr_type: string;
-  policy_rank: number;
-}
+import { Policy } from './types';
 
 export class AlPoliciesClientInstance {
